+++ src/StateReaderTaskEither.ts
@@ -2,12 +2,16 @@
  * @since 2.0.0
  */
 import { Alt4 } from './Alt'
+import { Applicative4 } from './Applicative'
+import { Apply4 } from './Apply'
 import { Bifunctor4 } from './Bifunctor'
 import { Either } from './Either'
 import { identity, pipe, Predicate, Refinement } from './function'
+import { Functor4 } from './Functor'
 import { IO } from './IO'
 import { IOEither } from './IOEither'
 import { Monad4 } from './Monad'
+import { MonadIO4 } from './MonadIO'
 import { MonadTask4 } from './MonadTask'
 import { MonadThrow4 } from './MonadThrow'
 import { Option } from './Option'
@@ -15,13 +19,8 @@ import { Reader } from './Reader'
 import { ReaderEither } from './ReaderEither'
 import * as RTE from './ReaderTaskEither'
 import { State } from './State'
-import * as StateT from './StateT'
 import { Task } from './Task'
 import { TaskEither } from './TaskEither'
-import { Functor4 } from './Functor'
-import { Apply4 } from './Apply'
-import { Applicative4 } from './Applicative'
-import { MonadIO4 } from './MonadIO'
 
 import ReaderTaskEither = RTE.ReaderTaskEither
 
@@ -55,18 +54,26 @@ export interface StateReaderTaskEither<S, R, E, A> {
  */
 export const evaluate: <S>(
   s: S
-) => <R = unknown, E = never, A = never>(ma: StateReaderTaskEither<S, R, E, A>) => ReaderTaskEither<R, E, A> =
-  /*#__PURE__*/
-  StateT.evaluate(RTE.monadReaderTaskEither)
+) => <R = unknown, E = never, A = never>(ma: StateReaderTaskEither<S, R, E, A>) => ReaderTaskEither<R, E, A> = (s) => (
+  fsa
+) =>
+  pipe(
+    fsa(s),
+    RTE.map(([a]) => a)
+  )
 
 /**
  * Run a computation in the `StateReaderTaskEither` monad discarding the result
  *
  * @since 3.0.0
  */
-export const execute: <S>(s: S) => <R, E, A>(ma: StateReaderTaskEither<S, R, E, A>) => ReaderTaskEither<R, E, S> =
-  /*#__PURE__*/
-  StateT.execute(RTE.monadReaderTaskEither)
+export const execute: <S>(s: S) => <R, E, A>(ma: StateReaderTaskEither<S, R, E, A>) => ReaderTaskEither<R, E, S> = (
+  s
+) => (fsa) =>
+  pipe(
+    fsa(s),
+    RTE.map(([_, s]) => s)
+  )
 
 /**
  * @since 2.0.0
@@ -78,9 +85,8 @@ export function left<S, R, E = never, A = never>(e: E): StateReaderTaskEither<S,
 /**
  * @since 2.0.0
  */
-export const right: <S, R, E = never, A = never>(a: A) => StateReaderTaskEither<S, R, E, A> =
-  /*#__PURE__*/
-  StateT.of(RTE.monadReaderTaskEither)
+export const right: <S, R, E = never, A = never>(a: A) => StateReaderTaskEither<S, R, E, A> = (a) => (s) =>
+  RTE.right([a, s])
 
 /**
  * @since 2.0.0
@@ -148,9 +154,9 @@ export function leftIO<S, R, E = never, A = never>(me: IO<E>): StateReaderTaskEi
 /**
  * @since 2.0.0
  */
-export const rightState: <S, R, E = never, A = never>(ma: State<S, A>) => StateReaderTaskEither<S, R, E, A> =
-  /*#__PURE__*/
-  StateT.fromState(RTE.monadReaderTaskEither)
+export const rightState: <S, R, E = never, A = never>(ma: State<S, A>) => StateReaderTaskEither<S, R, E, A> = (sa) => (
+  s
+) => RTE.right(sa(s))
 
 /**
  * @since 2.0.0
@@ -162,45 +168,44 @@ export function leftState<S, R, E = never, A = never>(me: State<S, E>): StateRea
 /**
  * @since 2.0.0
  */
-export const fromReaderTaskEither: <S, R, E, A>(ma: ReaderTaskEither<R, E, A>) => StateReaderTaskEither<S, R, E, A> =
-  /*#__PURE__*/
-  StateT.fromF(RTE.monadReaderTaskEither)
+export const fromReaderTaskEither: <S, R, E, A>(ma: ReaderTaskEither<R, E, A>) => StateReaderTaskEither<S, R, E, A> = (
+  fa
+) => (s) =>
+  pipe(
+    fa,
+    RTE.map((a) => [a, s])
+  )
 
 /**
  * Get the current state
  *
  * @since 2.0.0
  */
-export const get: <S, R, E = never>() => StateReaderTaskEither<S, R, E, S> =
-  /*#__PURE__*/
-  StateT.get(RTE.monadReaderTaskEither)
+export const get: <S, R, E = never>() => StateReaderTaskEither<S, R, E, S> = () => (s) => RTE.right([s, s])
 
 /**
  * Set the state
  *
  * @since 2.0.0
  */
-export const put: <S, R, E = never>(s: S) => StateReaderTaskEither<S, R, E, void> =
-  /*#__PURE__*/
-  StateT.put(RTE.monadReaderTaskEither)
+export const put: <S, R, E = never>(s: S) => StateReaderTaskEither<S, R, E, void> = (s) => () =>
+  RTE.right([undefined, s])
 
 /**
  * Modify the state by applying a function to the current state
  *
  * @since 2.0.0
  */
-export const modify: <S, R, E = never>(f: (s: S) => S) => StateReaderTaskEither<S, R, E, void> =
-  /*#__PURE__*/
-  StateT.modify(RTE.monadReaderTaskEither)
+export const modify: <S, R, E = never>(f: (s: S) => S) => StateReaderTaskEither<S, R, E, void> = (f) => (s) =>
+  RTE.right([undefined, f(s)])
 
 /**
  * Get a value which depends on the current state
  *
  * @since 2.0.0
  */
-export const gets: <S, R, E = never, A = never>(f: (s: S) => A) => StateReaderTaskEither<S, R, E, A> =
-  /*#__PURE__*/
-  StateT.gets(RTE.monadReaderTaskEither)
+export const gets: <S, R, E = never, A = never>(f: (s: S) => A) => StateReaderTaskEither<S, R, E, A> = (f) => (s) =>
+  RTE.right([f(s), s])
 
 /**
  * @since 2.4.0
@@ -294,9 +299,18 @@ export const alt: <S, R, E, A>(
  */
 export const ap: <S, R, E, A>(
   fa: StateReaderTaskEither<S, R, E, A>
-) => <B>(fab: StateReaderTaskEither<S, R, E, (a: A) => B>) => StateReaderTaskEither<S, R, E, B> =
-  /*#__PURE__*/
-  StateT.ap(RTE.monadReaderTaskEither)
+) => <B>(fab: StateReaderTaskEither<S, R, E, (a: A) => B>) => StateReaderTaskEither<S, R, E, B> = (fa) => (fab) => (
+  s1
+) =>
+  pipe(
+    fab(s1),
+    RTE.chain(([f, s2]) =>
+      pipe(
+        fa(s2),
+        RTE.map(([a, s3]) => [f(a), s3])
+      )
+    )
+  )
 
 /**
  * @since 2.0.0
@@ -339,9 +353,11 @@ export const bimap: <E, G, A, B>(
  */
 export const chain: <S, R, E, A, B>(
   f: (a: A) => StateReaderTaskEither<S, R, E, B>
-) => (ma: StateReaderTaskEither<S, R, E, A>) => StateReaderTaskEither<S, R, E, B> =
-  /*#__PURE__*/
-  StateT.chain(RTE.monadReaderTaskEither)
+) => (ma: StateReaderTaskEither<S, R, E, A>) => StateReaderTaskEither<S, R, E, B> = (f) => (ma) => (s1) =>
+  pipe(
+    ma(s1),
+    RTE.chain(([a, s2]) => f(a)(s2))
+  )
 
 /**
  * @since 2.0.0
@@ -405,9 +421,11 @@ export const flatten: <S, R, E, A>(
  */
 export const map: <A, B>(
   f: (a: A) => B
-) => <S, R, E>(fa: StateReaderTaskEither<S, R, E, A>) => StateReaderTaskEither<S, R, E, B> =
-  /*#__PURE__*/
-  StateT.map(RTE.monadReaderTaskEither)
+) => <S, R, E>(fa: StateReaderTaskEither<S, R, E, A>) => StateReaderTaskEither<S, R, E, B> = (f) => (fa) => (s1) =>
+  pipe(
+    fa(s1),
+    RTE.map(([a, s2]) => [f(a), s2])
+  )
 
 /**
  * @since 2.6.2