+++ src/index.ts
@@ -74,7 +74,6 @@ import * as semiring from './Semiring'
 import * as show from './Show'
 import * as state from './State'
 import * as stateReaderTaskEither from './StateReaderTaskEither'
-import * as stateT from './StateT'
 import * as store from './Store'
 import * as task from './Task'
 import * as taskEither from './TaskEither'
@@ -378,10 +377,6 @@ export {
    * @since 2.0.0
    */
   stateReaderTaskEither,
-  /**
-   * @since 2.0.0
-   */
-  stateT,
   /**
    * @since 2.0.0
    */